requirements.txt
aioredis==1.3.1
alabaster==0.7.12
amqp==2.6.1
APScheduler==3.7.0
asgiref==3.3.1
async-timeout==3.0.1
atomicwrites==1.4.0
attrs==20.3.0
autobahn==21.3.1
Automat==20.2.0
Babel==2.9.0
billiard==3.6.4.0
celery==4.4.7
certifi==2020.12.5
cffi==1.14.5
channels==3.0.3
channels-redis==3.2.0
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
colorama==0.4.4
constantly==15.1.0
cryptography==3.4.7
daphne==3.0.2
defusedxml==0.7.1
Django==3.1.7
psycopg2-binary==2.8.5
django-allauth==0.44.0
django-celery-beat==2.2.0
django-celery-results==2.0.1
django-ckeditor==6.0.0
django-js-asset==1.2.2
django-phone-field==1.8.1
django-redis==4.12.1
django-stubs==1.8.0
django-stubs-ext==0.2.0
django-timezone-field==4.1.2
dnspython==1.16.0
docutils==0.16
eventlet==0.30.2
factory-boy==3.2.0
Faker==8.1.1
flake8==3.9.0
flake8-import-order==0.18.1
flower==0.9.7
greenlet==1.0.0
hiredis==2.0.0
humanize==3.4.1
hyperlink==21.0.0
idna==2.10
imagesize==1.2.0
incremental==21.3.0
iniconfig==1.1.1
install==1.3.4
Jinja2==2.11.3
kombu==4.6.11
loguru==0.5.3
MarkupSafe==1.1.1
mccabe==0.6.1
msgpack==1.0.2
mypy==0.812
mypy-extensions==0.4.3
oauthlib==3.1.0
packaging==20.9
Pillow==8.1.2
pluggy==0.13.1
prometheus-client==0.8.0
prompt-toolkit==3.0.18
py==1.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
pycparser==2.20
pyflakes==2.3.0
Pygments==2.8.1
PyJWT==1.7.1
pyOpenSSL==20.0.1
pyparsing==2.4.7
pytest==6.2.2
pytest-django==4.2.0
python-crontab==2.5.1
python-dateutil==2.8.1
python-dotenv==0.17.0
python3-openid==3.2.0
pytz==2021.1
redis==3.5.3
requests==2.25.1
requests-oauthlib==1.3.0
service-identity==18.1.0
six==1.15.0
snowballstemmer==2.1.0
sorl-thumbnail==12.7.0
Sphinx==3.5.3
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
sqlparse==0.4.1
text-unidecode==1.3
toml==0.10.2
tornado==6.1
twilio==6.56.0
Twisted==21.2.0
txaio==21.2.1
typed-ast==1.4.3
typing-extensions==3.7.4.3
tzlocal==2.1
urllib3==1.26.4
vine==1.3.0
wcwidth==0.2.5
win32-setctime==1.0.3
zope.interface==5.4.0
sentry-sdk==1.0.0
djangorestframework==3.12.4
django-filter==2.4.0